requirements.txt
aioredis==1.2.0
amqp==2.5.1
asgi-redis==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
boto3==1.10.14
botocore==1.13.14
celery==4.3.0
certifi==2018.10.15
cffi==1.13.0
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.2.2
dateparser==0.7.2
dateutils==0.6.6
defusedxml==0.5.0
distro==1.4.0
dj-database-url==0.5.0
Django==2.2
django-admin-rangefilter==0.3.6
django-allauth==0.38.0
django-autocomplete-light==3.2.10
django-background-tasks==1.1.13
django-braces==1.13.0
django-celery-beat==1.5.0
django-celery-results==1.0.1
django-ckeditor==5.7.1
django-compat==1.0.15
django-cors-headers==2.4.0
django-countries==5.5
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-filter==2.1.0
django-froala-editor==2.8.5
django-jet==1.0.8
django-js-asset==1.2.2
django-model-utils==3.1.2
django-multiselectfield==0.1.8
django-polymorphic==2.1.2
django-rest-api==0.1.5
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-rest-framework-braces==0.3.4
django-rest-polymorphic==0.1.8
django-storages==1.7.1
django-timezone-field==3.1
djangorestframework==3.9.0
docutils==0.15.2
gunicorn==19.7.1
hiredis==0.2.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.7
importlib-metadata==0.23
incremental==17.5.0
isort==4.3.4
jeepney==0.4.1
jet-bridge-base==0.6.6
jet-django==0.8.9
jmespath==0.9.4
joblib==0.14.0
jsonfield==2.0.2
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.5
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lxml==4.4.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.5.6
msgpack-python==0.5.6
mysqlclient==1.4.4
netifaces==0.10.4
nltk==3.4.5
numpy==1.17.4
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
pbr==5.4.3
pexpect==4.2.1
Pillow==5.1.0
prompt-toolkit==2.0.9
protobuf==3.0.0
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycparser==2.19
pycrypto==2.6.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.1.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-crontab==2.3.9
python-dateutil==2.8.0
python-debian==0.1.32
python3-openid==3.1.0
pytz==2018.7
pyxdg==0.25
PyYAML==3.12
redis==3.3.0
regex==2019.11.1
reportlab==3.4.0
requests==2.20.1
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
rest-social-auth==2.2.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.2
SecretStorage==2.3.1
Serializer==0.2.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
SQLAlchemy==1.3.11
sqlparse==0.3.0
testresources==2.0.1
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.0
tzlocal==2.0.0
urllib3==1.24.3
vine==1.3.0
wadllib==1.3.2
wcwidth==0.1.7
whitenoise==3.3.1
wrapt==1.10.11
zipp==0.6.0
zope.interface==4.6.0
